+++ include/physics/kp_ray.hpp
@@ -39,9 +39,9 @@ namespace KalaPhysics::Physics
 		friend class KalaPhysics::Core::PhysicsWorld;
 	public:
 		//Create a new mask from multiple layers
-		static inline u64 MakeMaskFromLayers(initializer_list<u8> layers)
+		static inline u32 MakeMaskFromLayers(initializer_list<u8> layers)
 		{
-			u64 m = 0ULL;
+			u32 m = 0ULL;
 			for (u8 l : layers)
 			{
 				if (l < MAX_LAYERS) m |= (1ULL << l);
@@ -65,7 +65,7 @@ namespace KalaPhysics::Physics
 			f32 maxDistance = 0.0f);
 		
 		//Set mask directly
-		inline void SetMask(u64 m) { mask = m; }
+		inline void SetMask(u32 m) { mask = m; }
 		//Reset mask (no collisions)
 		inline void ClearMask() { mask = 0ULL; }
 		
@@ -74,8 +74,8 @@ namespace KalaPhysics::Physics
 		//Exclude layer by name
 		void RemoveLayerFromMask(const string& layer);
 		
-		inline u64 GetMask() const { return mask; }
+		inline u32 GetMask() const { return mask; }
 	private:
-		u64 mask = ~0ULL; //default - collide with everything
+		u32 mask = ~0ULL; //default - collide with everything
 	};
 }
